+++ spec/rails_helper.rb
@@ -11,6 +11,7 @@ require 'database_cleaner'
 require 'support/database_cleaner'
 require 'webmock/rspec'
 require 'support/vcr_setup'
+require 'capybara/rspec'
 WebMock.disable_net_connect!(allow_localhost: true)  
 # Add additional requires below this line. Rails is not loaded until this point!